requirements.txt
alembic==1.3.2
blinker==1.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Mail-SendGrid==0.3
Flask-Migrate==2.5.2
Flask-SendGrid==0.7.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-http-client==3.2.1
sendgrid==6.1.0
six==1.13.0
SQLAlchemy==1.3.12
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
gunicorn


sqlalchemy-s3sqlite
# installing stuff for docker inside of dockerfile
## neeeded for docker mysql
#pymysql
## needed for docker postgresql
## psycopg2-binary
